Cutscene.py and changes to MainMenu.py
I've added in a working cutscene file! I've also changed up MainMenu so that this goes through the cutscene.

What this file does is it allow us to tell a story and add a more premium feel to the game by making fade in and out transitions between scenes and the main menu etc. with the enter key and to increase game time and the anticipation

I've also added a fade out for the audio in the main menu so that it doesn't sound out of the place, feel free to change the images/add more images to the cutscene file if you wish to import your own however I'll be making images and a story for this, I've also commented everything accordingly as well